+++ CHANGES.md
@@ -2,6 +2,7 @@
 
 ### Version 0.9.6:
 - Added new option visibleDays, which states the days that are visible (defaults to all days, see documentation).
+- Fixed a fault that prevented the first event in a repeating series from being removed.
 - 
 
 ### Version 0.9.5: